requirements.txt
mysql_connector_repackaged==0.3.1
Pillow==10.1.0
Pillow==10.1.0
streamlit==1.28.2
streamlit_option_menu==0.3.6
